.claude/skills/input_structure/requirements.txt
pymupdf>=1.24.0
paddlepaddle>=2.6.0
paddleocr>=2.8.0
python-pptx>=0.6.23
python-docx>=1.1.0
pillow>=10.0.0
chardet>=5.0.0
